Integrate hooks into about
Needed a bit of local state for this component as I wanted the <p> tags
to display one at a time, based on a setInterval call. Was able to succ
essfully integrate 5 timers and their corresponding hooks to each of th
e <p> tags listing content about me. Most likely will not keep all of
it and will refactor in the future, as it is a bit busy on the screen,
but functions as expected for now.