Update project-page links
Added actual URLs for projects to projects-page. Used the pathname attr
ibute of the Link component as well as a target=_blank to have a new ta
b opened in the browser when a user clicks the project link. Updated st
ylings to remove underline on links. Appears to be working as expected.