Add contact page component, update styles
Added a contact-page component to provide a means to reach me for anyon
e using the site who may be so inclined to. Added 2 SVG images for the
github and linkedin links and imported them as react components. Added
and adjusted styles. Will be deploying this version, but will continue
improvements on it.