Expand testing suites
Added an ideas.txt file to keep track of my thoughts while working, and
added it to the .gitignore file. Expanded testing suites by clarifying
the App.test.js tests for rendering success. Added a second test file f
or HomePage component, which currently renders it shallowly and looks f
or a successful render of an h1 with class 'intro-title' as well as 2 p
tags with the class 'title-detail'. All tests run successfully, and fai
l as expected when altered.